requirements.txt
streamlit
requests
sentence-transformers
faiss-cpu
python-dotenv
numpy
